[#1599] Factor Session model back to single file
